Feat/#29: 꾸미기 모드 시작기능

+++ src/main/java/com/picpic/server/room/dto/ws/DecorateStickerTransformRequestDTO.java
@@ -4,7 +4,7 @@ public record DecorateStickerTransformRequestDTO(
         Long roomId,
         Long stickerInstanceId,
         Long stickerId,
-        Integer scale,
+        Double scale,
         Integer rotate
 ) {
 }
